buffetsmart/requirements.txt
django
gunicorn
djangorestframework==3.15.2
googletrans==4.0.0rc1
pillow==10.4.0
requests==2.32.3
psycopg2-binary==2.9.9
psycopg2==2.9.9
django-cors-headers
djangorestframework-simplejwt==5.3.1
jwt==1.3.1
pytz==2024.1
django-filter==24.3
APScheduler==3.10.4
SQLAlchemy==2.0.24
anyio==3.7.1
aiohttp==3.10.5
jwt==1.3.1
pillow==10.4.0
requests==2.32.3
uvicorn==0.25.0
PyJWT==2.9.0
deepl==1.19.1
django-tenants==3.7.0
deepl==1.19.1
drf-yasg==1.21.8
pip install django-rest-swagger==2.2.0
pandas==2.2.2
openpyxl==3.1.5